override redshift default
Redshift's default configuration (when using a config file) usually
prevents changes from being applied from the command line. Hence,
in this commit we kill the redshift process to make changed from
the command line.

Also, we get the current redshift's configuration as basis for the
changes that redshift-adjust will make.